# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Athenas API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/{person_id}/delete": {
            "delete": {
                "tags": [
                    "/{person_id}/delete"
                ],
                "summary": "Deletar pessoa pelo ID",
                "description": "Deletar pessoa pelo ID",
                "operationId": "49ddfdd3b708e9107770df5ad603b2b0",
                "parameters": [
                    {
                        "name": "person_id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Pessoa deletada com sucesso!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}